COMP: Fix compilation with VTK 7.1
When Slicer is built with the current VTK master (upcoming VTK 7.1), compilation
fails because of several deprecated headers and methods. This commit fixes
compilation with VTK 7.1 by updating the usage of the deprecated headers and
methods. Compatibility with the current version of VTK that Slicer uses is
maintained.

Specific changes accounted for include:

- Kitware/VTK@d9c5ca0
  (Mark legacy methods as "legacy".)

- Kitware/VTK@3ae7dd3
  (ENH: Remove use of include <vtksys/ios/*> and vtksys_ios::*)

- Kitware/VTK@eaf0f6a
  (ENH: Remove use of include <vtksys/stl/*> and vtksys_stl::*)


git-svn-id: http://svn.slicer.org/Slicer3/branches/Slicer4-EMSegment@17119 3bd1e089-480b-0410-8dfb-8563597acbee